A lot of debugging on the aggregator (see more)
Basically changed a lot of stuff so the aggregator can actually aggregate. Still need to verify the running average algorithm is correct by using different weights (I've only used the same weight which should always yield the same output as the input). Also need to re-serve the new model/weights when continuing to the next round